pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "statsmodels>=0.14.4",
    "streamlit-elements>=0.1.0",
    "streamlit>=1.45.0",
    "trafilatura>=2.0.0",
    "python-docx>=1.1.2",
    "pillow>=11.2.1",
    "google-generativeai>=0.8.5",
]
